requirements.txt
# Audio recording
sounddevice>=0.4.6
soundfile>=0.12.1

# Transcription (Whisper on Apple Silicon via MLX)
mlx-whisper>=0.1.0

# Speaker diarization
pyannote.audio>=3.1.0
torch>=2.1.0
torchaudio>=2.1.0

# Local LLM for summarization (Phi-3 on Apple Silicon)
mlx-lm>=0.10.0

# Encrypted database (for production — requires sqlcipher C library)
# pysqlcipher3>=1.2.0  # TODO: enable when ready for encryption

# PDF export
reportlab>=4.1.0

# UI
PyQt6>=6.6.0
